+++ test/array.spec.js
@@ -129,4 +129,19 @@ describe('Array Functions', () => {
     expect(toggleB(true)).to.deep.equal(['a', 'b', 'c', 'b'])
     expect(toggleB(false)).to.deep.equal(['a', 'c'])
   })
+  it('intersperse', () => {
+    expect(
+      F.intersperse(
+        (acc, i, xs) => (i === xs.length - 1 ? 'and finally' : 'and'),
+        [1, 2, 3]
+      )
+    ).to.deep.equal([1, 'and', 2, 'and finally', 3])
+    expect(F.intersperse('and', [1, 2, 3])).to.deep.equal([
+      1,
+      'and',
+      2,
+      'and',
+      3,
+    ])
+  })
 })